only try to remove file when it exists

+++ sources/libraries/cxxu/cxxu/utils.cpp
@@ -275,6 +275,10 @@ rmfile(const std::string& path)
     namespace fs = boost::filesystem;
     fs::path ppath(path);
 
+    if (!fs::exists(ppath)) {
+        return false;
+    }
+
     return fs::remove_all(ppath);
 }
